Update menu items and logic locations
