+++ app/src/main/java/com/playground/navigationwithtabs/ui/TabContentFragment.kt
@@ -37,8 +37,6 @@ class TabContentFragment : Fragment(), View.OnClickListener {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        setHasOptionsMenu(true)
-
         type = arguments?.getString("type", "") ?: ""
     }
 
@@ -67,26 +65,6 @@ class TabContentFragment : Fragment(), View.OnClickListener {
                 listAdapter.setTasks(value)
             }
         }
-    }
-
-    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
-        inflater.inflate(R.menu.app_menu, menu)
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        when (item.itemId) {
-            R.id.menu_item_add -> addTab()
-            R.id.menu_item_delete -> deleteTabs()
-        }
-        return true
-    }
-
-    private fun addTab() {
-        requireView().findNavController().navigate(R.id.addTabFragment)
-    }
-
-    private fun deleteTabs() {
-        requireView().findNavController().navigate(R.id.deleteCategoriesFragment)
     }
 
     override fun onClick(v: View) {
