+++ app/src/main/res/layout/fragment_view_pager.xml
@@ -29,6 +29,7 @@
         </androidx.appcompat.widget.Toolbar>
 
         <com.google.android.material.tabs.TabLayout
+            style="@style/StandardTabLayoutStyle"
             android:id="@+id/tab_layout"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
