+++ app/src/main/res/values-night/styles.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <style name="TextAppearance.MaterialComponents.Subtitle1.ListItem">
+        <item name="android:textColor">@color/white</item>
+    </style>
+    <style name="TextAppearance.MaterialComponents.Subtitle1.ListItem.Completed">
+        <item name="android:textColor">@color/grey_light</item>
+    </style>
+</resources>
